+++ misc/update_DESCRIPTION_NEWS.R
@@ -4,6 +4,7 @@
 # devtools::install_github("Dschaykib/newsmd")
 # install.packages("desc")
 library(newsmd)
+library(desc)
 
 # update roxygen
 roxygen2::roxygenise()
@@ -329,7 +330,8 @@ my_news$add_bullet(c("removing ubuntu with R version 4.0 in the test setup",
                      "changing lintr options"))
 
 
-# adjusting test setup ----------------------------------------------------
+
+# backwards compability ---------------------------------------------------
 
 my_desc$bump_version("patch")
 my_news$add_version(my_desc$get_version())
@@ -350,14 +352,21 @@ my_news$add_bullet(c("data.tree",
 my_desc$del_dep("data.tree", type = desc::dep_types[1])
 
 
-# rebuilding print_fs -----------------------------------------------------
+
+# removing dependencies ---------------------------------------------------
 
 my_desc$bump_version("patch")
 my_news$add_version(my_desc$get_version())
-my_news$add_subtitle("removing dependencies")
+my_news$add_subtitle("removing direct dependencies")
 my_news$add_bullet(c("readr", "magrittr"))
 my_desc$del_dep("readr", type = desc::dep_types[1])
 my_desc$del_dep("magrittr", type = desc::dep_types[1])
+
+my_news$add_subtitle("adding dependencies")
+my_news$add_bullet(c("utils", "grid"))
+my_desc$set_dep("utils", type = desc::dep_types[1])
+my_desc$set_dep("grid", type = desc::dep_types[1])
+
 
 # save everything ---------------------------------------------------------